Bug in learnNeuron + bunch of improvements + MNIST example code in Lib.hs (#1)
Fixes a bug in learnNeuron: probably code I wrote while sleepy (also VS Code doesn't warn me of unused parameters or variables. That would have raised red flags).

After this I should provide a better structure for different executables on different examples, and write better tests. Then maybe we can go back to the idea of using vectors, though they'll have to be mutable probably, which is terrible. But something should probably be done about the implicit matrix transposition that happens in `errorL`.

Only after all that we could move on to do a better, perhaps monadic representation of a network, plus finishing the book http://neuralnetworksanddeeplearning.com/chap6.html with a basic convolutional network. That'd be a wrap